+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroupsSettings/MSFT_AADGroupsSettings.psm1
@@ -42,9 +42,21 @@ function Get-TargetResource
         [System.String]
         $Ensure = 'Present',
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
-        $GlobalAdminAccount
+        $GlobalAdminAccount,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
 
     Write-Verbose -Message "Getting configuration of AzureAD Groups Naming Policy"
@@ -55,8 +67,22 @@ function Get-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    Test-MSCloudLogin -CloudCredential $GlobalAdminAccount `
-        -Platform AzureAD
+    $ConnectionMode = $null
+    if (-not [String]::IsNullOrEmpty($ApplicationId) -and `
+        -not [String]::IsNullOrEmpty($TenantId) -and `
+        -not [String]::IsNullOrEmpty($CertificateThumbprint))
+    {
+        Write-Verbose -Message "Connecting to AzureAD using ApplicationId {$ApplicationId}"
+        Test-MSCloudLogin -Platform AzureAD -ApplicationId $ApplicationId -TenantId $TenantId -CertificateThumbprint $CertificateThumbprint
+        $ConnectionMode = "ServicePrincipal"
+    }
+    else
+    {
+        Write-Verbose -Message "Connecting to AzureAD using Credentials"
+        Test-MSCloudLogin -CloudCredential $GlobalAdminAccount `
+            -Platform AzureAD
+        $ConnectionMode = "Credential"
+    }
 
     $Policy = Get-AzureADDirectorySetting | Where-Object -FilterScript {$_.DisplayName -eq "Group.Unified"}
 
@@ -77,15 +103,18 @@ function Get-TargetResource
         }
         $result = @{
             IsSingleInstance               = 'Yes'
-            EnableGroupCreation            = $Policy["EnableGroupCreation"]
-            AllowGuestsToBeGroupOwner      = $Policy["AllowGuestsToBeGroupOwner"]
-            AllowGuestsToAccessGroups      = $Policy["AllowGuestsToAccessGroups"]
+            EnableGroupCreation            = [Boolean]$Policy["EnableGroupCreation"]
+            AllowGuestsToBeGroupOwner      = [Boolean]$Policy["AllowGuestsToBeGroupOwner"]
+            AllowGuestsToAccessGroups      = [Boolean]$Policy["AllowGuestsToAccessGroups"]
             GuestUsageGuidelinesUrl        = $Policy["GuestUsageGuidelinesUrl"]
             GroupCreationAllowedGroupName  = $AllowedGroupName
-            AllowToAddGuests               = $Policy["AllowToAddGuests"]
+            AllowToAddGuests               = [Boolean]$Policy["AllowToAddGuests"]
             UsageGuidelinesUrl             = $Policy["UsageGuidelinesUrl"]
             Ensure                         = "Present"
             GlobalAdminAccount             = $GlobalAdminAccount
+            ApplicationId                  = $ApplicationId
+            TenantId                       = $TenantId
+            CertificateThumbprint          = $CertificateThumbprint
         }
 
         Write-Verbose -Message "Get-TargetResource Result: `n $(Convert-M365DscHashtableToString -Hashtable $result)"
@@ -136,9 +165,21 @@ function Set-TargetResource
         [System.String]
         $Ensure = 'Present',
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
-        $GlobalAdminAccount
+        $GlobalAdminAccount,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
 
     Write-Verbose -Message "Setting configuration of Azure AD Groups Settings"
@@ -148,9 +189,6 @@ function Set-TargetResource
     $data.Add("Method", $MyInvocation.MyCommand)
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
-
-    Test-MSCloudLogin -CloudCredential $GlobalAdminAccount `
-        -Platform AzureAD
 
     $currentPolicy = Get-TargetResource @PSBoundParameters
 
@@ -234,9 +272,21 @@ function Test-TargetResource
         [System.String]
         $Ensure = 'Present',
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
-        $GlobalAdminAccount
+        $GlobalAdminAccount,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
 
     Write-Verbose -Message "Testing configuration of AzureAD Groups Settings"
@@ -264,9 +314,21 @@ function Export-TargetResource
     [OutputType([System.String])]
     param
     (
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
-        $GlobalAdminAccount
+        $GlobalAdminAccount,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
     $InformationPreference = 'Continue'
     #region Telemetry
@@ -275,18 +337,51 @@ function Export-TargetResource
     $data.Add("Method", $MyInvocation.MyCommand)
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
-
-    $content = ''
-    $params = @{
-        GlobalAdminAccount = $GlobalAdminAccount
-        IsSingleInstance   = 'Yes'
+    $ConnectionMode = $null
+    if (-not [String]::IsNullOrEmpty($ApplicationId) -and `
+        -not [String]::IsNullOrEmpty($TenantId) -and `
+        -not [String]::IsNullOrEmpty($CertificateThumbprint))
+    {
+        $params = @{
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
+            IsSingleInstance      = 'Yes'
+        }
+        $ConnectionMode = "ServicePrincipal"
     }
+    else
+    {
+        $params = @{
+            GlobalAdminAccount = $GlobalAdminAccount
+            IsSingleInstance   = 'Yes'
+        }
+        $ConnectionMode = "Credential"
+    }
+    $content = ''
     $result = Get-TargetResource @params
-    $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
+    if ($ConnectionMode -eq 'Credential')
+    {
+        $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
+        $result.Remove("ApplicationId")
+        $result.Remove("TenantId")
+        $result.Remove("CertificateThumbprint")
+    }
+    else
+    {
+        $result.Remove("GlobalAdminAccount")
+    }
     $content += "        AADGroupsSettings " + (New-GUID).ToString() + "`r`n"
     $content += "        {`r`n"
     $currentDSCBlock = Get-DSCBlock -Params $result -ModulePath $PSScriptRoot
-    $content += Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
+    if ($ConnectionMode -eq 'Credential')
+    {
+        $content += Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
+    }
+    else
+    {
+        $content += $currentDSCBlock
+    }
     $content += "        }`r`n"
 
     return $content